feat: Layout component
