+++ src/App.css
@@ -1,125 +1,10 @@
+/* Importar estilos base */
 @import "./styles/variables.css";
 @import "./styles/base.css";
-
-:root {
-  --shadow-light: 0 2px 4px rgba(0, 0, 0, 0.1);
-  --shadow-medium: 0 4px 8px rgba(0, 0, 0, 0.12);
-}
 
 /* Override específico para el fondo de la app */
 body {
   background-color: var(--color-bg);
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-}
-
-.app {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-}
-
-.app-header {
-  background: var(--color-bg);
-  border-bottom: 1px solid var(--color-border);
-  padding: var(--spacing-md) var(--spacing-lg);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  box-shadow: var(--shadow-light);
-  position: sticky;
-  top: 0;
-  z-index: 100;
-}
-
-.app-title {
-  font-size: var(--font-size-lg);
-  font-weight: 700;
-  margin: 0;
-  color: var(--color-accent);
-  cursor: pointer;
-  transition: opacity 0.2s ease;
-}
-
-.app-title:hover {
-  opacity: 0.8;
-}
-
-.loading-indicator {
-  width: 12px;
-  height: 12px;
-  background: var(--color-accent);
-  border-radius: 50%;
-  animation: pulse 1s ease-in-out infinite;
-}
-
-@keyframes pulse {
-  0%,
-  100% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.3;
-  }
-}
-
-.app-main {
-  flex: 1;
-  padding: var(--spacing-lg);
-  max-width: 1400px;
-  margin: 0 auto;
-  width: 100%;
-}
-
-.error-message {
-  background: #fee;
-  color: #c33;
-  padding: var(--spacing-md);
-  border-radius: var(--radius-card);
-  border: 1px solid #fcc;
-  margin-bottom: var(--spacing-md);
-}
-
-.results-count {
-  background: var(--color-bg);
-  display: inline-block;
-  padding: var(--spacing-sm) var(--spacing-md);
-  border-radius: calc(var(--radius-card) * 1.25);
-  font-size: var(--font-size-sm);
-  font-weight: 600;
-  color: var(--color-text-muted);
-  margin-bottom: calc(var(--spacing-lg) * 0.75);
-  box-shadow: var(--shadow-light);
-}
-
-.podcast-grid {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr); /* 4 columnas fijas */
-  gap: calc(var(--spacing-lg) * 0.75);
-  /* list-style y padding ya están en base.css */
-  margin: 0;
-  /* Padding top para el primer row de imágenes flotantes más grandes */
-  padding-top: calc(130px / 2);
-}
-
-@media (min-width: 768px) {
-  .podcast-grid {
-    gap: var(--spacing-lg);
-  }
-
-  .app-main {
-    padding: calc(var(--spacing-lg) * 1.5);
-  }
-}
-
-@media (max-width: 767px) {
-  .podcast-grid {
-    grid-template-columns: repeat(2, 1fr); /* 2 columnas en móvil */
-  }
-}
-
-@media (min-width: 1024px) {
-  .podcast-grid {
-    grid-template-columns: repeat(5, 1fr); /* 5 columnas en desktop grande */
-  }
 }
