+++ src/components/Layout.css
@@ -1,0 +1,68 @@
+@import "../styles/variables.css";
+
+.layout {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  background-color: var(--color-bg);
+}
+
+.layout-header {
+  background: var(--color-bg);
+  border-bottom: 1px solid var(--color-border);
+  padding: var(--spacing-md) var(--spacing-lg);
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  position: sticky;
+  top: 0;
+  z-index: 1000;
+}
+
+.layout-title {
+  font-size: var(--font-size-lg);
+  font-weight: 700;
+  margin: 0;
+  color: var(--color-accent);
+  cursor: pointer;
+  transition: opacity 0.2s ease;
+}
+
+.layout-title:hover {
+  opacity: 0.8;
+}
+
+.layout-loading-indicator {
+  width: 12px;
+  height: 12px;
+  background: var(--color-accent);
+  border-radius: 50%;
+  animation: pulse 1s ease-in-out infinite;
+}
+
+@keyframes pulse {
+  0%,
+  100% {
+    opacity: 1;
+  }
+  50% {
+    opacity: 0.3;
+  }
+}
+
+.layout-main {
+  flex: 1;
+  padding: var(--spacing-lg);
+  max-width: 1920px;
+  margin: 0 auto;
+  width: 100%;
+  /* Asegurar que el contenido no se salga por arriba */
+  overflow-x: hidden;
+}
+
+@media (min-width: 768px) {
+  .layout-main {
+    padding: calc(var(--spacing-lg) * 1.5);
+  }
+}
