+++ src/components/PodcastCard.css
@@ -1,5 +1,3 @@
-/* src/components/PodcastCard.css */
-/* Importar variables (base.css ya se importa en App.css) */
 @import "../styles/variables.css";
 
 .podcast-card {
@@ -46,13 +44,11 @@
   height: 130px;
   border-radius: 50%; /* Imagen circular */
   object-fit: cover;
-  border: 4px solid var(--color-bg); /* Borde blanco para destacar */
   /* Sin sombra en la imagen para evitar sombra superior */
 }
 
 .podcast-card__content {
   text-align: center;
-  /* Sin margin-top porque ya no tenemos image-container con margin-bottom */
 }
 
 .podcast-card__title {
@@ -79,5 +75,3 @@
   -webkit-box-orient: vertical;
   overflow: hidden;
 }
-
-/* Responsive grid - esto irá en App.css ya está implementado */
